chore: PHPStan Level 5 cleanup — add type annotations, new model methods, request classes
- Add AutoScalingHistory model methods (getActionLabel, isSuccess, etc.)
- Add AutoScalingRule model constants and methods
- Add Organization model with members/servers counts
- Add AdminRole permissions (billing discount codes, webhooks create, server presets)
- Add AlertSettingsRequest and ModeSettingsRequest validation
- Add OrganizationTransformer for client API
- Fix return type annotations across 35 files
- Add phpstan.neon configuration

+++ app/Http/Controllers/Auth/Modules/DiscordLoginController.php
@@ -85,7 +85,7 @@ class DiscordLoginController extends AbstractLoginController
      */
     public function randStr(int $length = 10): string
     {
-        return substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', ceil($length / strlen($x)))), 1, $length);
+        return substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', (int) ceil($length / strlen($x)))), 1, $length);
     }
 }
